requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.118
botocore==1.34.118
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.4
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
jmespath==1.0.1
multidict==6.0.5
mysql-connector==2.2.9
mysqlclient==2.2.3
packaging==24.0
phonenumbers==8.13.38
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.3.1
twilio==9.1.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
